+++ CHANGELOG.md
@@ -19,6 +19,7 @@ The releases of Vieb aim to follow [semantic versioning](https://semver.org).
 - Pacman archive as a new included build target (for Arch Linux)
 - Arm64 architecture builds for most Linux and Windows releases
 - Add windowtitle setting to optionally include the page title and/or url in Vieb's window title
+- Recursive insert mode mappings are now possible
 
 ### Changed